+++ src/perplexity_encoder.h
@@ -1,4 +1,4 @@
 #pragma once
 #include "encoder_interface.h"
 
-void perplexity_encoder_new(EncoderInterface* enc, float* perplexity);
+void perplexity_encoder_new(EncoderInterface* enc, uint64_t* perplexity);
